Update Code and fixing minor bugs

+++ application/sources/app/game/in_the_depth_game/itd_game_border.h
@@ -20,6 +20,6 @@
 
 extern uint16_t itd_game_score;
 extern uint16_t itd_game_time;
-extern uint8_t itd_game_heart;
+extern uint16_t itd_game_heart;
 
 #endif /*_ITD_GAME_BORDER_H_*/
